Fix TestCaddyStartStop
Move this test to caddymain to fix import problems
Minor fixes to bypass locks and race conditions
