+++ caddy/caddymain/run_test.go
@@ -15,8 +15,12 @@
 package caddymain
 
 import (
+	"net/http"
 	"runtime"
 	"testing"
+	"time"
+
+	"github.com/mholt/caddy"
 )
 
 func TestSetCPU(t *testing.T) {
@@ -55,5 +59,31 @@ func TestSetCPU(t *testing.T) {
 		}
 		// teardown
 		runtime.GOMAXPROCS(currentCPU)
+	}
+}
+
+func TestCaddyStartStop(t *testing.T) {
+	caddyfile := "localhost:1984"
+
+	for i := 0; i < 2; i++ {
+		testInst, err := caddy.Start(caddy.CaddyfileInput{Contents: []byte(caddyfile)})
+		if err != nil {
+			t.Fatalf("Error starting, iteration %d: %v", i, err)
+		}
+
+		client := http.Client{
+			Timeout: time.Duration(1 * time.Second),
+		}
+		resp, err := client.Get("http://localhost:1984")
+		if err != nil {
+			t.Fatalf("Expected GET request to succeed (iteration %d), but it failed: %v", i, err)
+		}
+		resp.Body.Close()
+
+		err = testInst.Stop()
+		if err != nil {
+			t.Fatalf("Error stopping, iteration %d: %v", i, err)
+		}
+		time.Sleep(1)
 	}
 }
